+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/main/java/com/tencent/cloud/polaris/circuitbreaker/config/PolarisCircuitBreakerBootstrapConfiguration.java
@@ -33,10 +33,9 @@ import org.springframework.context.annotation.Configuration;
  *
  * @author lepdou 2022-03-29
  */
-@ConditionalOnPolarisEnabled
-@ConditionalOnProperty(value = "spring.cloud.polaris.circuitbreaker.enabled", havingValue = "true",
-		matchIfMissing = true)
 @Configuration(proxyBeanMethods = false)
+@ConditionalOnPolarisEnabled
+@ConditionalOnProperty(value = "spring.cloud.polaris.circuitbreaker.enabled", havingValue = "true", matchIfMissing = true)
 public class PolarisCircuitBreakerBootstrapConfiguration {
 
 	@Bean
@@ -66,7 +65,5 @@ public class PolarisCircuitBreakerBootstrapConfiguration {
 		public int getOrder() {
 			return ContextConstant.ModifierOrder.CIRCUIT_BREAKER_ORDER;
 		}
-
 	}
-
 }