+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/main/java/com/tencent/cloud/polaris/circuitbreaker/feign/PolarisFeignBlockingLoadBalancerClient.java
@@ -34,5 +34,4 @@ public class PolarisFeignBlockingLoadBalancerClient extends FeignBlockingLoadBal
 			LoadBalancerClientFactory loadBalancerClientFactory) {
 		super(delegate, loadBalancerClient, loadBalancerClientFactory);
 	}
-
 }