+++ spring-cloud-starter-tencent-polaris-router/src/main/java/com/tencent/cloud/polaris/router/config/PolarisRuleBasedRouterProperties.java
@@ -44,5 +44,4 @@ public class PolarisRuleBasedRouterProperties {
 				"enabled=" + enabled +
 				'}';
 	}
-
 }