+++ spring-cloud-tencent-polaris-context/src/main/java/com/tencent/cloud/polaris/context/config/PolarisContextPostConfiguration.java
@@ -29,7 +29,7 @@ import org.springframework.context.annotation.Configuration;
  * Post-initialization operations after the application initialization phase is completed.
  *@author lepdou 2022-06-28
  */
-@Configuration
+@Configuration(proxyBeanMethods = false)
 public class PolarisContextPostConfiguration {
 
 	@Bean