+++ spring-cloud-tencent-polaris-context/src/main/java/com/tencent/cloud/polaris/context/config/PolarisContextProperties.java
@@ -132,5 +132,4 @@ public class PolarisContextProperties {
 	public void setService(String service) {
 		this.service = service;
 	}
-
 }