+++ spring-cloud-tencent-polaris-loadbalancer/src/test/java/com/tencent/cloud/polaris/loadbalancer/config/PolarisRouterAutoConfigurationTest.java
@@ -84,5 +84,4 @@ public class PolarisRouterAutoConfigurationTest {
 	static class PolarisLoadBalancerTest {
 
 	}
-
 }